+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/server/FederatedHMSHandler.java
@@ -1509,8 +1509,15 @@ class FederatedHMSHandler extends FacebookBase implements CloseableIHMSHandler {
 
   @Override
   @Loggable(value = Loggable.DEBUG, skipResult = true, name = INVOCATION_LOG_NAME)
-  public GetAllFunctionsResponse get_all_functions() throws MetaException, TException {
-    return getPrimaryClient().get_all_functions();
+  public GetAllFunctionsResponse get_all_functions() throws TException {
+    DatabaseMapping mapping = databaseMappingService.primaryDatabaseMapping();
+    GetAllFunctionsResponse allFunctions = mapping.getClient().get_all_functions();
+    if(allFunctions.getFunctions() != null) {
+      for (Function function : allFunctions.getFunctions()) {
+        mapping.transformOutboundFunction(function);
+      }
+    }
+    return allFunctions;
   }
 
   @Override