+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## TBD
+### Changed
+* Allow primary metastore to have a prefix. See [#152](https://github.com/HotelsDotCom/waggle-dance/issues/152).
+
 ## [3.2.0] - 2019-03-27
 ### Added
 * Configurable `latency` for each metastore in a Waggle Dance configuration.